+++ tang-extension/tang-mybatis-flex/src/main/kotlin/com/tang/extension/mybatisflex/utils/PageUtils.kt
@@ -23,6 +23,9 @@ object PageUtils {
     @JvmStatic
     fun <T> page(baseMapper: BaseMapper<T>, queryWrapper: QueryWrapper): PageResult {
         val pageDomain = PageSupport.buildPageRequest()
+        if (pageDomain.orderBy.isNotBlank()) {
+            queryWrapper.orderBy(pageDomain.orderBy, pageDomain.isAscBoolean)
+        }
         val paginate = baseMapper.paginate(pageDomain.pageNum, pageDomain.pageSize, queryWrapper)
         return PageResult(paginate.records, paginate.totalRow)
     }